requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.4.0
Django==4.2.6
django-browser-reload==1.12.0
django-environ==0.11.2
django-filter==23.3
django-htmx==1.17.0
django-tailwind==3.6.0
idna==3.4
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
psycopg==3.1.12
psycopg2==2.9.9
Pygments==2.16.1
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7